MPT-17825: Refactor products commands into per-command modules
Mirror the price-lists structure: replace the monolithic
`cli/core/products/app.py` with a package `cli/core/products/app/`
that hosts one module per Typer command.

- `export.py`: introduce `ProductExporter` holding the run-wide state
  (account container, out dir). Adopt the temp-file + atomic-replace
  pattern so a failed export does not destroy the previous workbook
  (`with_stem` keeps the `.xlsx` suffix that openpyxl requires). Loop
  the related service factories instead of inlining 10 export calls.
- `list_products.py`: minor cleanup; share the MPT client across pages,
  inline the offset computation, return early instead of carrying a
  `has_pages` flag.
- `sync.py`: introduce `ProductSyncer` exposing `sync`, with private
  `_confirm`, `_create_parameter_collections`, `_run_create` and
  `_run_update` helpers. Merge the previous create/update confirm
  branches into one `_confirm` method. Drive `_run_update` from a
  tuple of `(status_message, service_factory)` steps to avoid a
  hand-rolled chain of 10 status updates.
- `app.py`: aggregate the three sub-app Typer instances, matching the
  price-lists pattern.
- Mirror the test structure: `tests/cli/core/products/app/` with
  `test_export.py`, `test_list_products.py`, `test_sync.py`. Update
  the conftest mock paths to point at the new modules. The
  `create_product`/`update_product` helper-function tests become
  `ProductSyncer._run_create`/`_run_update` method tests.

Drop the WPS210/WPS213/WPS231 entry for `cli/core/products/app.py`
from `[tool.flake8] per-file-ignores`. `_run_create` retains a
narrow inline `# noqa: WPS210,WPS213` because the create-product
flow genuinely owns several intermediate collections.

No behavior changes; ruff/flake8/mypy/pytest all green (362 passed,
100% coverage on export.py and list_products.py, 99% on sync.py).
