+++ cli/core/products/__init__.py
@@ -1,3 +1,3 @@
-from cli.core.products.app import app
+from cli.core.products.app.app import app
 
 __all__ = ["app"]
