--- tests/cli/core/products/test_app.py
@@ -1,484 +1,0 @@
-from pathlib import Path
-
-import pytest
-from cli.core.models import DataCollectionModel
-from cli.core.mpt.models import Meta, Product
-from cli.core.products import app as product_app
-from cli.core.products.app import create_product, update_product
-from cli.core.services.service_result import ServiceResult
-from cli.core.stats import ProductStatsCollector
-from mpt_api_client.models import Collection, Model, Pagination
-from mpt_api_client.models import Meta as ClientMeta
-from openpyxl.reader.excel import load_workbook
-from rich.status import Status
-from typer.testing import CliRunner
-
-runner = CliRunner()
-
-
-def test_list_products(active_vendor_account, mocker, mpt_products_response):
-    mocker.patch("cli.core.products.app.get_active_account", return_value=active_vendor_account)
-    mocker.patch(
-        "cli.core.products.app.get_products",
-        return_value=(
-            Meta(limit=10, offset=0, total=2),
-            [
-                Product.model_validate(product_data)
-                for product_data in mpt_products_response["data"]
-            ],
-        ),
-    )
-
-    result = runner.invoke(product_app, ["list"])
-
-    assert result.exit_code == 0, result.stdout
-    assert mpt_products_response["data"][0]["id"] in result.stdout
-
-
-def test_list_products_with_query_and_paging(active_vendor_account, mocker, mpt_products_response):
-    mocker.patch("cli.core.products.app.get_active_account", return_value=active_vendor_account)
-    mocker.patch(
-        "cli.core.products.app.get_products",
-        return_value=(
-            Meta(limit=20, offset=0, total=2),
-            [
-                Product.model_validate(product_data)
-                for product_data in mpt_products_response["data"]
-            ],
-        ),
-    )
-
-    result = runner.invoke(
-        product_app,
-        ["list", "--page", "20", "--query", "eq(product.id,'PRD-1234')"],
-    )
-
-    assert result.exit_code == 0, result.stdout
-    assert mpt_products_response["data"][0]["id"] in result.stdout
-
-
-def test_sync_not_valid_definition(mocker, product_container_mock):
-    mocker.patch.object(
-        product_container_mock.product_service(),
-        "validate_definition",
-        return_value=ServiceResult(success=False, model=None, stats=mocker.Mock()),
-    )
-
-    result = runner.invoke(
-        product_app,
-        ["sync", "--dry-run", "some-file"],
-    )
-
-    assert result.exit_code == 3, result.stdout
-
-
-def test_sync_with_dry_run(mocker, product_container_mock):
-    create_mock = mocker.patch("cli.core.products.app.create_product")
-    update_mock = mocker.patch("cli.core.products.app.update_product")
-
-    result = runner.invoke(
-        product_app,
-        ["sync", "--dry-run", "fake_file.xlsx"],
-    )
-
-    assert result.exit_code == 0, result.stdout
-    assert "Product definition" in result.stdout
-    create_mock.assert_not_called()
-    update_mock.assert_not_called()
-
-
-def test_sync_product_update(mocker, product_container_mock, product_data_from_dict):
-    validate_definition_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "validate_definition",
-        return_value=ServiceResult(success=True, model=None, stats=mocker.Mock()),
-    )
-    retrieve_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "retrieve",
-        return_value=ServiceResult(success=True, model=product_data_from_dict, stats=mocker.Mock()),
-    )
-    update_product_mock = mocker.patch("cli.core.products.app.update_product")
-
-    result = runner.invoke(
-        product_app,
-        ["sync", "fake_file.xlsx"],
-        input="y\n",
-    )
-
-    assert result.exit_code == 0, result.stdout
-    assert "Do you want to update product" in result.stdout
-    validate_definition_mock.assert_called_once()
-    retrieve_mock.assert_called_once()
-    update_product_mock.assert_called_once()
-
-
-def test_sync_product_update_error(mocker, product_container_mock, product_data_from_dict):
-    mocker.patch.object(ProductStatsCollector, "has_errors", new=True)
-    validate_definition_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "validate_definition",
-        return_value=ServiceResult(success=True, model=None, stats=mocker.Mock()),
-    )
-    retrieve_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "retrieve",
-        return_value=ServiceResult(success=True, model=product_data_from_dict, stats=mocker.Mock()),
-    )
-    update_product_mock = mocker.patch("cli.core.products.app.update_product")
-
-    result = runner.invoke(
-        product_app,
-        ["sync", "fake_file.xlsx"],
-        input="y\n",
-    )
-
-    assert result.exit_code == 3, result.stdout
-    validate_definition_mock.assert_called_once()
-    retrieve_mock.assert_called_once()
-    update_product_mock.assert_called_once()
-
-
-def test_sync_product_force_create(
-    mocker, product_container_mock, product_new_file, product_data_from_dict
-):
-    mocker.patch.object(ProductStatsCollector, "has_errors", new=False)
-    render_mock = mocker.patch("cli.core.products.app.stats_table_renderer.render", return_value="")
-    validate_definition_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "validate_definition",
-        return_value=ServiceResult(success=True, model=None, stats=mocker.Mock()),
-    )
-    retrieve_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "retrieve",
-        return_value=ServiceResult(success=True, model=product_data_from_dict, stats=mocker.Mock()),
-    )
-    create_product_mock = mocker.patch("cli.core.products.app.create_product")
-
-    result = runner.invoke(
-        product_app,
-        ["sync", "--force-create", str(product_new_file)],
-        input="y\n",
-    )
-
-    assert result.exit_code == 0, result.stdout
-    assert "Do you want to create new" in result.stdout
-    validate_definition_mock.assert_called_once()
-    retrieve_mock.assert_called_once()
-    create_product_mock.assert_called_once()
-    render_mock.assert_called_once()
-
-
-def test_sync_product_no_product(mocker, product_new_file, product_container_mock):
-    mocker.patch.object(ProductStatsCollector, "has_errors", new=False)
-    render_mock = mocker.patch("cli.core.products.app.stats_table_renderer.render", return_value="")
-    validate_definition_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "validate_definition",
-        return_value=ServiceResult(success=True, model=None, stats=mocker.Mock()),
-    )
-    retrieve_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "retrieve",
-        return_value=ServiceResult(success=True, model=None, stats=mocker.Mock()),
-    )
-    create_mock = mocker.patch("cli.core.products.app.create_product")
-
-    result = runner.invoke(
-        product_app,
-        ["sync", str(product_new_file)],
-        input="y\n",
-    )
-
-    assert result.exit_code == 0, result.stdout
-    assert "Do you want to create new product for account" in result.stdout
-    validate_definition_mock.assert_called_once()
-    retrieve_mock.assert_called_once()
-    create_mock.assert_called_once()
-    render_mock.assert_called_once()
-
-
-def test_create_product(
-    mocker,
-    product_new_file,
-    product_data_from_dict,
-    product_related_data,
-    product_container_mock,
-):
-    create_product_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "create",
-        return_value=ServiceResult(success=True, model=product_data_from_dict, stats=mocker.Mock()),
-    )
-    create_item_group_mock = mocker.patch.object(
-        product_container_mock.item_group_service(),
-        "create",
-        return_value=ServiceResult(
-            success=True,
-            model=None,
-            collection=DataCollectionModel(
-                collection={"fake_id": product_related_data["item_group"]}
-            ),
-            stats=mocker.Mock(),
-        ),
-    )
-    create_parameter_group_mock = mocker.patch.object(
-        product_container_mock.parameter_group_service(),
-        "create",
-        return_value=ServiceResult(
-            success=True,
-            model=None,
-            collection=DataCollectionModel(
-                collection={"fake_id": product_related_data["parameter_group"]}
-            ),
-            stats=mocker.Mock(),
-        ),
-    )
-    create_parameter_service_mock = mocker.patch.object(
-        product_container_mock.agreement_parameters_service(),
-        "create",
-        return_value=ServiceResult(
-            success=True,
-            model=None,
-            collection=DataCollectionModel(
-                collection={"fake_id": product_related_data["parameters"]}
-            ),
-            stats=mocker.Mock(),
-        ),
-    )
-    set_new_parameter_mock = mocker.patch.object(
-        product_container_mock.agreement_parameters_service(), "set_new_parameter_group"
-    )
-    create_template_service_mock = mocker.patch.object(
-        product_container_mock.template_service(),
-        "create",
-        return_value=ServiceResult(
-            success=True, model=product_related_data["template"], stats=mocker.Mock()
-        ),
-    )
-    set_new_parameter_template_mock = mocker.patch.object(
-        product_container_mock.template_service(), "set_new_parameter_group"
-    )
-    create_item_service_mock = mocker.patch.object(
-        product_container_mock.item_service(),
-        "create",
-        return_value=ServiceResult(
-            success=True,
-            model=product_related_data["item"],
-            stats=mocker.Mock(),
-        ),
-    )
-    set_new_item_groups_mock = mocker.patch.object(
-        product_container_mock.item_service(), "set_new_item_groups"
-    )
-    add_collection_spy = mocker.spy(DataCollectionModel, "add")
-
-    create_product(product_container_mock, mocker.Mock(spec=Status))  # act
-
-    create_product_mock.assert_called_once()
-    assert product_container_mock.resource_id() == product_data_from_dict.id
-    create_item_group_mock.assert_called_once()
-    create_parameter_group_mock.assert_called_once()
-    create_parameter_service_mock.assert_called_once()
-    set_new_parameter_mock.assert_called_once()
-    assert add_collection_spy.call_count == 4
-    create_template_service_mock.assert_called_once()
-    set_new_parameter_template_mock.assert_called_once()
-    create_item_service_mock.assert_called_once()
-    set_new_item_groups_mock.assert_called_once()
-
-
-def test_create_product_error(mocker, product_container_mock):
-    stats = product_container_mock.stats()
-    create_product_mock = mocker.patch.object(
-        product_container_mock.product_service(),
-        "create",
-        return_value=ServiceResult(success=False, model=None, stats=stats),
-    )
-    create_item_group_spy = mocker.spy(product_container_mock.item_service(), "create")
-
-    create_product(product_container_mock, mocker.Mock(spec=Status))  # act
-
-    create_product_mock.assert_called_once()
-    create_item_group_spy.assert_not_called()
-
-
-def test_export_product_account_not_allowed(account_container_mock, active_vendor_account):
-    account_container_mock.account.override(active_vendor_account)
-
-    result = runner.invoke(product_app, ["export", "fake_id"])
-
-    assert result.exit_code == 4, result.stdout
-
-
-def test_export_product_error_exporting_product(mocker, product_container_mock):
-    product_container_mock.stats.override(mocker.Mock(spec=ProductStatsCollector, has_errors=True))
-
-    result = runner.invoke(product_app, ["export", "fake_id"], input="y\n")
-
-    assert result.exit_code == 3, result.stdout
-
-
-def test_export_product_overwrites_existing_files(
-    mocker,
-    tmp_path,
-    product_container_mock,
-):
-    exists_mock = mocker.patch.object(Path, "exists", return_value=True)
-    unlink_mock = mocker.patch.object(Path, "unlink", return_value=True)
-    product_id = "PRD-1234"
-    tmp_file = tmp_path / f"{product_id}.xlsx"
-    tmp_file.touch()
-
-    result = runner.invoke(
-        product_app,
-        ["export", product_id, "--out", str(tmp_file)],
-        input="y\ny\n",
-    )
-
-    assert result.exit_code == 0, result.stdout
-    exists_mock.assert_called_once()
-    assert tmp_file.exists()
-    unlink_mock.assert_called_once()
-    product_container_mock.product_service().export.assert_called_once()
-
-
-def test_update_product(mocker, product_container_mock):
-    update_product(product_container_mock, mocker.Mock(spec=Status))  # act
-
-    product_container_mock.item_service().update.assert_called_once()
-    product_container_mock.item_group_service().update.assert_called_once()
-    product_container_mock.template_service().update.assert_called_once()
-    product_container_mock.parameter_group_service().update.assert_called_once()
-    product_container_mock.agreement_parameters_service().update.assert_called_once()
-    product_container_mock.asset_parameters_service().update.assert_called_once()
-    product_container_mock.request_parameters_service().update.assert_called_once()
-    product_container_mock.subscription_parameters_service().update.assert_called_once()
-
-
-def make_collection(mocker, data_list):
-    collection = mocker.MagicMock(spec=Collection)
-    collection.meta = mocker.MagicMock(spec=ClientMeta)
-    collection.meta.pagination = mocker.MagicMock(spec=Pagination)
-    collection.meta.pagination.limit = 100
-    collection.meta.pagination.offset = 0
-    collection.meta.pagination.total = len(data_list)
-    resources = []
-    for resource_data in data_list:
-        resource = mocker.MagicMock(spec=Model)
-        resource.to_dict.return_value = resource_data
-        resources.append(resource)
-    collection.resources = resources
-    return collection
-
-
-@pytest.mark.integration
-def test_export_product(
-    tmp_path,
-    mocker,
-    account_container_mock,
-    mpt_product_data,
-    product_related_data,
-):
-    mock_api_client = account_container_mock.api_mpt_client()
-    product_resource = mocker.MagicMock(spec=Model)
-    product_resource.to_dict.return_value = mpt_product_data
-    mock_api_client.catalog.products.get.return_value = product_resource
-    mock_items = mock_api_client.catalog.items
-    mock_items_select = mock_items.filter.return_value.select.return_value
-    mock_items_select.fetch_page.return_value = make_collection(
-        mocker, [product_related_data["mpt_item"]]
-    )
-    item_groups = mock_api_client.catalog.products.item_groups.return_value
-    item_groups.select.return_value.fetch_page.return_value = make_collection(
-        mocker, [product_related_data["mpt_item_group"]]
-    )
-    parameter_groups = mock_api_client.catalog.products.parameter_groups.return_value
-    parameter_groups.select.return_value.fetch_page.return_value = make_collection(
-        mocker, [product_related_data["mpt_parameter_group"]]
-    )
-    params_service = mock_api_client.catalog.products.parameters.return_value
-    params_service_select = params_service.filter.return_value.select.return_value
-    params_service_select.fetch_page.side_effect = [
-        make_collection(mocker, [product_related_data["mpt_agreement_parameter"]]),
-        make_collection(mocker, [product_related_data["mpt_asset_parameter"]]),
-        make_collection(mocker, [product_related_data["mpt_item_parameter"]]),
-        make_collection(mocker, [product_related_data["mpt_request_parameter"]]),
-        make_collection(mocker, [product_related_data["mpt_subscription_parameter"]]),
-    ]
-    products_templates = mock_api_client.catalog.products.templates.return_value
-    products_templates.select.return_value.fetch_page.return_value = make_collection(
-        mocker, [product_related_data["mpt_template"]]
-    )
-    product_id = "PRD-0232-2541"
-
-    result = runner.invoke(
-        product_app, ["export", product_id, "--out", str(tmp_path)], input="y\ny\n"
-    )
-
-    assert result.exit_code == 0, result.stdout
-    product_path = tmp_path / f"{product_id}.xlsx"
-    assert product_path.exists()
-    wb = load_workbook(product_path)
-    expected_sheets = [
-        "General",
-        "Settings",
-        "Items",
-        "Items Groups",
-        "Parameters Groups",
-        "Agreements Parameters",
-        "Assets Parameters",
-        "Item Parameters",
-        "Request Parameters",
-        "Subscription Parameters",
-        "Templates",
-    ]
-    assert wb.sheetnames == expected_sheets
-    general_sheet = wb["General"]
-    assert general_sheet.max_row == 12
-    assert general_sheet["A1"].value == "General Information"
-    assert general_sheet["A2"].value == "Product ID"
-    assert general_sheet["B2"].value == product_id
-    settings_sheet = wb["Settings"]
-    assert settings_sheet.max_row == 12
-    assert settings_sheet["A1"].value == "Setting"
-    assert settings_sheet["A2"].value == "Change order validation (draft)"
-    assert settings_sheet["C2"].value == "Enabled"
-    items_sheet = wb["Items"]
-    assert items_sheet.max_row == 2
-    assert items_sheet["A1"].value == "ID"
-    assert items_sheet["A2"].value == "ITM-0232-2541-0001"
-    items_groups_sheet = wb["Items Groups"]
-    assert items_groups_sheet.max_row == 2
-    assert items_groups_sheet["A1"].value == "ID"
-    assert items_groups_sheet["A2"].value == "IGR-0232-2541-0001"
-    parameters_groups_sheet = wb["Parameters Groups"]
-    assert parameters_groups_sheet.max_row == 2
-    assert parameters_groups_sheet["A1"].value == "ID"
-    assert parameters_groups_sheet["A2"].value == "PGR-0232-2541-0002"
-    agreements_params_sheet = wb["Agreements Parameters"]
-    assert agreements_params_sheet.max_row == 2
-    assert agreements_params_sheet["A1"].value == "ID"
-    assert agreements_params_sheet["A2"].value == "PAR-0232-2541-0001"
-    agreements_params_sheet = wb["Assets Parameters"]
-    assert agreements_params_sheet.max_row == 2
-    assert agreements_params_sheet["A1"].value == "ID"
-    assert agreements_params_sheet["A2"].value == "PAR-0232-2541-0027"
-    item_params_sheet = wb["Item Parameters"]
-    assert item_params_sheet.max_row == 2
-    assert item_params_sheet["A1"].value == "ID"
-    assert item_params_sheet["A2"].value == "PAR-0232-2541-0022"
-    request_params_sheet = wb["Request Parameters"]
-    assert request_params_sheet.max_row == 2
-    assert request_params_sheet["A1"].value == "ID"
-    assert request_params_sheet["A2"].value == "PAR-0232-2541-0012"
-    subscription_parameter_sheet = wb["Subscription Parameters"]
-    assert subscription_parameter_sheet.max_row == 2
-    assert subscription_parameter_sheet["A1"].value == "ID"
-    assert subscription_parameter_sheet["A2"].value == "PAR-0232-2541-0023"
-    templates_sheet = wb["Templates"]
-    assert templates_sheet.max_row == 2
-    assert templates_sheet["A1"].value == "ID"
-    assert templates_sheet["A2"].value == "TPL-0232-2541-0005"
-    wb.close()
